+++ src/EntityGraphQL/Schema/BaseSchemaTypeWithFields.cs
@@ -111,7 +111,7 @@ namespace EntityGraphQL.Schema
         }
 
         public abstract ISchemaType AddAllBaseTypes();
-        public abstract ISchemaType AddBaseType<TClrType>();
-        public abstract ISchemaType AddBaseType(string name);
+        public abstract ISchemaType Implements<TClrType>(bool addTypeIfNotInSchema = true, bool addAllFieldsOnAddedType = true);
+        public abstract ISchemaType Implements(string typeName);
     }
 }