+++ src/EntityGraphQL/Schema/MutationType.cs
@@ -153,12 +153,11 @@ public class MutationSchemaType : BaseSchemaTypeWithFields<MutationField>
     {
         throw new Exception("Cannot add base types to a mutation");
     }
-    public override ISchemaType AddBaseType<TClrType>()
+    public override ISchemaType Implements<TClrType>(bool addTypeIfNotInSchema = true, bool addAllFieldsOnAddedType = true)
     {
         throw new Exception("Cannot add base types to a mutation");
-
     }
-    public override ISchemaType AddBaseType(string name)
+    public override ISchemaType Implements(string typeName)
     {
         throw new Exception("Cannot add base types to a mutation");
     }