+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Breaking changes:
 - Cleaned up `SchemaType` constructors - using `GqlTypeEnum` instead of many boolean flags
 - Removed obsolete `SchemaProvider.AddInheritedType<TBaseType>`
 - Removed the instance parameter from `AddMutationsFrom` and friends. Mutation "controllers" are now always created per request like an asp.net controller. Use DI for any constructor parameters
+- Renamed `ISchemaType.AddBaseType` to `ISchemaType.Implements` to align with GraphQL language
 
 Changes: